Rename block.py File
The chomper/block.py file has been renamed to chomper/block_ref.py to
make block.py file contain a Block class and other modular
functionality. The block_ref.py is kept as the reference implementation.
To be used with testing to verify expected functionality is maintained.

WARNING: This breaks functionality. Other files and scripts need to be
refactored to keep reference functionality.